+++ sql_app/schemas.py
@@ -3,8 +3,8 @@ from pydantic import BaseModel
 
 class ItemCoord(BaseModel):
     id: int
-    lat: int
-    lon: int
+    lat: float
+    lon: float
     height: int
     rainny: int
     nitrogen: int
